requirements.txt
flask
picamera2
imageio
rawpy
opencv-python
pillow